# backend/serverPanels/swagger-output.json
{
    "swagger": "2.0",
    "info": {
        "title": "Nile Maret-Place Api",
        "description": "Description",
        "version": "1.0.0"
    },
    "host": "localhost:3001",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/cart": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "query",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/cartWebsite": {
            "put": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "customer_id": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/addProductcart": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "customer_id": {
                                    "example": "any"
                                },
                                "vendor_id": {
                                    "example": "any"
                                },
                                "product_uniqueid": {
                                    "example": "any"
                                },
                                "category": {
                                    "example": "any"
                                },
                                "subcategory": {
                                    "example": "any"
                                },
                                "quantity": {
                                    "example": "any"
                                },
                                "variantlabel": {
                                    "example": "any"
                                },
                                "mrp": {
                                    "example": "any"
                                },
                                "sellingprice": {
                                    "example": "any"
                                },
                                "c_symbol": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/removeProductFromCart": {
            "delete": {
                "description": "",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "subcategory",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "product_uniqueid",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "label",
                        "in": "query",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/removeProductFromCartWebsite": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "category": {
                                    "example": "any"
                                },
                                "subcategory": {
                                    "example": "any"
                                },
                                "product_uniqueid": {
                                    "example": "any"
                                },
                                "customer_id": {
                                    "example": "any"
                                },
                                "label": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    }
}